requirements.txt
langchain>=0.2.11
langchain-community>=0.2.10
langchain-openai>=0.1.20
chromadb>=0.5.5
pypdf>=4.2.0
sentence-transformers>=3.0.1
streamlit>=1.36.0
python-dotenv>=1.0.1
